changes made to enable the hook kea-dhcp-hooks-UTD
Dockerfile - redid Dockerfile to reduce the layers and add
kea-dhcp-hooks-UTD to it

kea-conf.yml - implemented the change to use the hook

TunnelIPv6Interface.cpp / netif-mgmt.c - changed hard coded prefix
length to 32

+++ external-servers/kea/start_kea.sh
@@ -1,7 +1,10 @@
 #!/bin/bash
 
-rm kea-conf.json
-yq . < kea-conf.yml > kea-conf.json
+# rm kea-conf.json
+# yq . < kea-conf.yml > kea-conf.json
 
-docker build --progress=plain -t alex-kea .
+if ! docker build --progress=plain -t alex-kea .; then
+    echo failed to build docker
+    exit 1
+fi
 docker run --rm --net=host alex-kea
